Saves deepcopy of input and outputs for logging
This is meant to "snapshot" the input and outputs as they might contains
mutable objects that can be altered after spans have been created.

A common examples is with the OpenAI integration where the messages list is
often modified when multiple LLM calls are made one after another with the
same history.